requirements.txt
fastapi
uvicorn
pandas
feedgen
feedparser
jinja2
pymongo
bs4
python-dateutil
